Remove output folder and make change to create directory if needed

+++ R/functions/hpc_localize_fn.R
@@ -1,6 +1,5 @@
 ## HPC localize
 hpc_localize_fn <- function(file_f,
-                            output_folder,
                             grid_points_folder,
                             log_dist_RSSI_mdl,
                             tz = tz,
@@ -251,14 +250,19 @@ hpc_localize_fn <- function(file_f,
                         "/ml_localized/",
                         file_f))
     
-    ## Save lists as R data file
-    saveRDS(tag_loc_est,
-            outfile)
+    ## Create tag directory if needed
+    dir <- paste(head(unlist(strsplit(outfile,"/")),-1),collapse="/")
+    if (!dir.exists(dir)) dir.create(dir)
+      
+      ## Save lists as R data file
+      saveRDS(tag_loc_est,
+              outfile)
     
     cat("\n Finished file:", file_f_ab, "-", length(unique(dets_p$t_ind)), "intervals localized", 
         "after", round(as.numeric(difftime(Sys.time(), start_time, units = "mins")), 1), "minutes \n")
     
-  } else {
+  }
+  else {
     
     cat("\n No intervals to localize, skipped file:", file_f_ab)
